+++ src/Core/Lifecycle/BindingEvidenceGate.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace GravityPresentationProfiles\Core\Lifecycle;
-
-interface BindingEvidenceGate {
-    public function allowsBinding( $binding );
-}