+++ src/Core/Lifecycle/StateStore.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace GravityPresentationProfiles\Core\Lifecycle;
-
-interface StateStore {
-    public function load();
-
-    public function commit( $expected_revision, $next_state );
-}