+++ src/Core/Portable/ContractViolation.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace GravityPresentationProfiles\Core\Portable;
-
-use InvalidArgumentException;
-
-final class ContractViolation extends InvalidArgumentException {}